Change pom.xml in main folder (scribejava). Added print to console System.out.println to help debugging. So far, getting clock information works but xjc doesn't seem to unpack xsd files right.
The lines in TradeKing.java are shifted down a bit but I didn't make any change other than in main() and get().

Some constructors are now broken because of new scribejava. These are in TradeKingForeman.java

I kept netty and async-http-client the same so I can focus on fixing scribejava.